Bridge4

+++ Bridgeit/DataStructurePrograms/src/com/bridgeit/programs/NumberOfBinarySearchTree.java
@@ -1,3 +1,12 @@
+/**
+ * Purpose: taking test cases as user input and taking total number of nodes per test cases
+ * and finding the total number of arrangements for the entered nodes
+ * 
+ * @author Kumar Shubham
+ * @since 06-04-2018
+ *
+ */
+
 package com.bridgeit.programs;
 
 import com.bridgeit.utility.Utility;
@@ -6,12 +15,12 @@ public class NumberOfBinarySearchTree {
 	public static void main(String[] args) {
 		Utility utility = new Utility();
 		System.out.println("Enter the total number of testcases");
-		int testCases = utility.inputInteger();
+		double testCases = utility.inputInteger();
 		
 		while (testCases > 0) {
 			System.out.println("Enter total number of nodes");
-			long totalNodes = utility.inputInteger();
-			long totalTree = Utility.numberOfBinarySearchTree(totalNodes);
+			double totalNodes = utility.inputInteger();
+			double totalTree = Utility.numberOfBinarySearchTree(totalNodes);
 			System.out.println(totalTree);
 			testCases--;
 		}
